IT-3317: Install dependencies as rstudio user (#2)
Previously the R package dependencies were installed while running as root. This made the packages unavailable to the rstudio user which is the user used by the RStudio notebook. This PR updates the Dockerfile to switch to rstudio while doing package installation and creating the Python venv, which Reticulate will use.

Also installs boto3 which wasn't getting done correctly in the previous version.

Also updates build-push-action to v5.